+++ api/routes/profile.route.js
@@ -1,0 +1,9 @@
+import { Router } from "express";
+import auth from "../middleware/auth.js";
+import { updateProfile } from "../controllers/profile.controller.js";
+
+const router = Router();
+
+router.put("/profile", auth, updateProfile);
+
+export default router;
