+++ api/server.js
@@ -1,8 +1,6 @@
 import http from "http";
 import api from "./api.js";
 import "./config/database.js";
-import dotenv from "dotenv";
-dotenv.config();
 
 const server = http.createServer(api);
